refactor: rename tasks-provider to tasks-context
Replaced `tasks-provider` with `tasks-context` across tasks components to ensure consistency with the established naming conventions. Updated all related imports, exports, file paths, and references accordingly.
