+++ apps/nextjs/src/components/features/tasks/hooks/use-tasks-can-delete-task-section.ts
@@ -1,4 +1,4 @@
-import { useTasksContext } from '@/components/features/tasks/tasks-provider/tasks-provider';
+import { useTasksContext } from '@/components/features/tasks/tasks-provider/tasks-context';
 import { useTasksTaskSection } from './use-tasks-task-section';
 import { useTasksTaskSectionIds } from './use-tasks-task-section-ids';
