+++ apps/nextjs/src/components/features/tasks/tasks-calendar/tasks-calendar-list-item/hooks/use-list-item-style.ts
@@ -1,5 +1,5 @@
 import { useMemo } from 'react';
-import { useTasksContext } from '@/components/features/tasks/tasks-provider/tasks-provider';
+import { useTasksContext } from '@/components/features/tasks/tasks-provider/tasks-context';
 import type { FlexProps } from '@/components/ui/flex';
 import type { TextProps } from '@/components/ui/text';
 import { dateFns } from '@/shared/date-fns';
