+++ apps/nextjs/src/components/features/tasks/tasks-calendar/tasks-calendar-list-item/info/info.tsx
@@ -1,5 +1,5 @@
 import { memo } from 'react';
-import { useTasksContext } from '@/components/features/tasks/tasks-provider/tasks-provider';
+import { useTasksContext } from '@/components/features/tasks/tasks-provider/tasks-context';
 import { Flex, type FlexProps } from '@/components/ui/flex';
 import { ProjectDueInfo } from './project-due-info';
