+++ apps/nextjs/src/components/features/tasks/tasks-list/tasks-list-cells/tasks-name/expand-icon/expand-icon.tsx
@@ -1,5 +1,5 @@
 import { memo } from 'react';
-import { useTasksContext } from '@/components/features/tasks/tasks-provider/tasks-provider';
+import { useTasksContext } from '@/components/features/tasks/tasks-provider/tasks-context';
 import { Container } from './container';
 import { Empty } from './empty';
