+++ apps/nextjs/src/components/features/tasks/tasks-provider/index.ts
@@ -1,1 +1,1 @@
-export * from './tasks-provider';
+export * from './tasks-context';
